requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.110.0
uvicorn==0.29.0
3dLitematica